outbound_calls/requirements.txt
elevenlabs-mcp>=0.2.0
fastapi==0.103.1
uvicorn==0.23.2
websockets>=11.0.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.3.0
typing-extensions>=4.5.0
aiohttp>=3.8.0
python-multipart==0.0.6 